Fix missing random seed bug for LSTM model
When the seed on the command line was set to 0 this would not result
in the seed actually being set in the code. This theoretically
requires rerunning all experiments, but we simply do not have time for
this. It would only slightly change the results of one out of ten
runs, so the differences are negligible.